% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alloy.R
\docType{data}
\name{alloy}
\alias{alloy}
\title{Alloy data.}
\format{
A data frame with 5 rows and 2 variables:
\describe{
  \item{AnnealtTime}{Time to anneal}
  \item{PassPotent}{Postent}

}
}
\source{
\url{https://www.routledge.com/Statistics-for-Engineering-and-the-Sciences-Sixth-Edition/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
alloy
}
\description{
This is a short data set from MS.
}
\details{
This is a standard data set from the course MATH 4753 taken from the data sets provided by the text book.
}
\keyword{datasets}
