% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag.R
\docType{data}
\name{flag}
\alias{flag}
\title{The FLAG data set.}
\format{
A data frame with 235 rows and 10 variables:
\describe{
  \item{COST}{response}
  \item{DOTEST}{see MS page 627}
  \item{STATUS}{see MS page 627}

   }
}
\source{
\url{https://www.routledge.com/Statistics-for-Engineering-and-the-Sciences-Sixth-Edition/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
flag
}
\description{
A dataset containing variables associated with cost (y) of road construction.
}
\details{
This is a standard data set from the course MATH 4753 taken from the data sets provided by the text book.
}
\keyword{datasets}
