% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftc2.R
\docType{data}
\name{ftc2}
\alias{ftc2}
\title{ftc2 data set.}
\format{
A data frame with 25 rows and 4 variables:
\describe{
  \item{CO}{response}
  \item{WEIGHT}{weight of tobacco}
  \item{NICOTINE}{Nicotine in tobacco}
  \item{TAR}{Tar in tobacco}
   }
}
\source{
\url{https://www.routledge.com/Statistics-for-Engineering-and-the-Sciences-Sixth-Edition/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
ftc2
}
\description{
A dataset containing variables associated with CO emmission for smokers
}
\details{
This is a standard data set from the course MATH 4753 taken from the data sets provided by the text book.
}
\keyword{datasets}
