% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irr.R
\docType{data}
\name{irr}
\alias{irr}
\title{Irrigation data.}
\format{
A data set with 5 ros and 2 columns
\describe{
  \item{STRATEGY}{Categorical 5 levels}
  \item{FREQ}{Frequency }
  ...
}
}
\source{
\url{https://www.routledge.com/Statistics-for-Engineering-and-the-Sciences/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
irr
}
\description{
A data set from MS giving opinions of farmers to different irrigation strategies
}
\keyword{datasets}
