% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mile3.R
\docType{data}
\name{mile3}
\alias{mile3}
\title{Three mile island disaster.}
\format{
A data set with 6 rows and 3 columns
\describe{
  \item{EVAC}{Yes or No, 2 levels}
  \item{DISTANCE}{distance in 3 levels }
  \item{NUMBER}{Freqency}
  ...
}
}
\source{
\url{https://www.routledge.com/Statistics-for-Engineering-and-the-Sciences/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
mile3
}
\description{
A data set from MS giving attitudes to evacuation from the three mile island at various distances from the nuclear plant
}
\keyword{datasets}
