% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mynewt.R
\name{mynewt}
\alias{mynewt}
\title{Basic Newton Raphson}
\usage{
mynewt(x0, delta = 0.001, f, fdash)
}
\arguments{
\item{x0}{initial value}

\item{delta}{increment in x}

\item{f}{the function to find the derivative of}

\item{fdash}{the first derivative}
}
\value{
a plot with the segments of the NR technique
}
\description{
Basic Newton Raphson
}
\examples{
\dontrun{mynewt(10,f=function(x) x^2-5*x +6, fdash= function(x) 2*x -5)}
}
