% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcwidgets.R
\name{rcwidgets}
\alias{rcwidgets}
\title{How to control one widget from another}
\usage{
rcwidgets()
}
\value{
Interactive browser with two widgets
}
\description{
How to control one widget from another
}
\examples{
\dontrun{rcwidgets()}
}
